feat(daemon): split process-exit monitor by OS — unblock Linux build (#478)
monitor.go used kqueue (EVFILT_PROC NOTE_EXIT) with no build tag, so
`GOOS=linux go build ./...` failed to compile — step 0 of the
cross-platform observation work.

Split the pidMonitor (outbound.ProcessWatcher) into three build-tagged
implementations behind one NewMonitor constructor and a shared exitHandler
type:
  - monitor_darwin.go — the existing kqueue impl, unchanged.
  - monitor_linux.go  — pidfd_open(2) + poll(2) (Linux 5.3+, no privileges).
  - monitor_other.go  — no-op stub; sessions still observed, PIDs reaped by
                        the periodic scanner. Lets the !darwin && !linux
                        target compile (Windows seam).

`GOOS=linux go build ./...` now compiles on amd64 and arm64; the
processlifecycle stub builds for windows. Darwin behaviour is unchanged —
full `go test ./core/... -race` stays green.
